We now generate a .gitignore for each .cvsignore file.
We previously autogenerated .cvsignore files to cause (some of) the source
code files that we autogenerate to be ignored by the CVS version control
program. Now we also generate .gitignore files (whose contents are the same
as the .cvsignore files) so that the git version control program ignores the
same files.