Added a commented out version of external.lib.dir for in-project jars.
So if the uncommented one is replaced with the commented out one then (at
least in theory) the project should build when all of the external JAR files
that it depends on (xalan2.jar, etc.) are in the project's lib/external
subdirectory instead of some system-dependent directory outside of the
project.

Some of the libraries that this project depends on have been around for a
while and may not remain easily accessible forever, so one may want to have
local copies of them. In particular, jdom1.jar is from a legacy version of
the project.