Replaced using the Xalan C++ executable with the Java JAR file.
While it seemed like a good idea to use the newer C++ executable version of
Xalan, it resulted in some difficult-to-diagnose problems. Part of that was
due to the fact that in addition to some targets in build.xml,
SourceCodeGenerator.java and its subclasses also generate source code using
XSL transforms/stylesheets, and the latter is written to use the Java Xalan
library. But without the Xalan Java library installed the Java classes in
javax.xml.transform picked some other Transformer that didn't play at all
well with my XSL files, resulting in invalid generated code and
NullPointerExceptions occurring when processing the 'section-titles'
templates in both semantics.xsl and validity-constraints.xsl.