Improved the conversion of whitespace.
We used to be more concerned with maintaining aspects of the original
document like line breaks and blank lines, but while that's suitable for
converting documents that just contain data, it could mess up some of the
finer whitespace-related details of documents that are more like (X)HTML
documents. So now we normalize most whitespace, and try to ensure that we
don't add any where there was none or leave none where there was some.

Also moved the trimLeading() method to TextUtilities, renaming it to
trimLeadingWhitespace() in the process.