Fixed the result of converting document types.
The "internal subset" part of a document type can contain values that aren't
valid in Atria attributes, so outputNonEmptyAttribute() can't be used to
output it, and its value can contain newlines, so its value can extend past
the end of the comment line that we currently convert/translate document
types into. Since we do just translate them into a comment we just omit the
internal subset part from the result of the translation (though this may not
be a permanent solution).